[WinForms] Fixed get SelectedNode after the handle is destroyed
If you put TreeView on the Form and display the Form via ShowDialog, and then select any other node and then close the Form, treeView.SelectedNode will return the first node of the tree,  but should return the last one selected. This is because SelectedNode when the form is already destroyed returns pre_selected_node, which is initialized only when the TreeView is created.
Solution is to keep pre_selected_node in actual state after selection.